rtrlib: solve permission right with user change
when capabilities are not used in a system, the calling system should
have the user rights to change of netns(). in some daemons, it is usual
to start as root, then change the owner of the daemon to have less
privileges. Then, when some operations need some privileges, a quick
change of owner is operated. rtrlib has the vrf init API that is
expanded to provide 2 user ids, one which is the initial userid with all
the supposed rights, and the second which is the daemonized user. Like
that, when rtrlib will want to do some privileged operations to change
netns, then seteuid() will be called appropriately.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>